+++ stonecrop/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log - @stonecrop/stonecrop
 
+## 0.30.0
+Thu, 27 Aug 2026 12:30:32 GMT
+
+### Patches
+
+- Return records from collectRecordPayload in the shape the store holds, leaving inline link reduction to the adapter
+
 ## 0.29.0
 Tue, 25 Aug 2026 08:42:39 GMT